Add MJPG always-on streaming.
Previously, streams from the servers were only accessible through the communication framework, and only after being enabled with a ControlMessage. Now, servers are *always* streaming, and can instantly switch over to image capture mode when a stills ControlMessage is received. The stream is exposed at http://<server-ip>:8080/?action=stream and can be easily be embedded in a web UI. The method used to get stream imagery in this commit is somewhat fragile and crashes on occasion, but should be resolved soon.